Move hasheddan to emeritus maintainers
Serving as a Crossplane maintainer has been a true honor. Moving
forward, I will not be able to commit the time and focus that is
expected of a maintainer, and other folks have stepped up to own the
areas that I have previously been leading. I am excited to step aside
and watch them continue to grow and improve the project.

Thank you for allowing me to be a leader in this community.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>